requirements.txt
flask==3.1.1
flask-cors==6.0.1
chromadb==0.4.15
sentence-transformers>=2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
chardet==5.2.0
psutil==5.9.6
torch>=2.2.0
numpy>=1.24.0
pytest==7.4.2
huggingface-hub>=0.20.0
